перенес функции в utils, переенные в constans
